Crate rs_stats

Functions

average

average gets the number expressing the central or typical value in a set of data

cummulative_distrib

CDF return the CDF using the mean and the standard deviation given https://en.wikipedia.org/wiki/Cumulative_distribution_function#Definition

erf

erf returns the error function integrated between zero and x. Error function.

erfc

erfc returns the error function integrated between x and infinity called complementary error function

normal_cummulative_distrib

CDF return the CDF for the zscore given https://en.wikipedia.org/wiki/Cumulative_distribution_function#Definition

normal_probability_density

normal_probability_density return the PDF with z already normalized https://en.wikipedia.org/wiki/Probability_density_function

probability_density

probability_density normalize x using the mean and the standard deviation and return the PDF https://en.wikipedia.org/wiki/Probability_density_function

std_dev

std_dev return the standard deviation, the square root of the variance

std_err

std_err is the standard error, represnting the standard deviation of its distribution

variance

variance us the mean of the sum of all square deviation

zscore

the zscore represente the distance from the mean in stddev